Extracted cluster execution logic to a new class
